backend/requirements.txt
cffi==1.15.1
charset-normalizer==3.2.0
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.13.0
django-filter==23.1
django-import-export==3.2.0
django-templated-mail==1.1.1
psycopg2-binary==2.9.3
djoser==2.1.0
flake8==6.0.0
webcolors==1.11.1
python-decouple==3.5
drf-extra-fields==3.4.0
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
gunicorn==20.1.0
